requirements.txt
annotated-types==0.7.0
anyio==4.6.1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
